+++ frontend/src/components/MyForge/tabs/MachineTypes.tsx
@@ -9,6 +9,8 @@ import { Cross2Icon, PlusIcon } from "@radix-ui/react-icons";
 import * as Dialog from '@radix-ui/react-dialog';
 import ResourceSlots from './ResourceSlots';
 import { OmniAPI } from 'src/apis/OmniAPI';
+import { UserPermission } from 'src/enums';
+import useAuth from '../../Auth/useAuth';
 
 interface aemenuprops {
     isDialogOpen: boolean;
@@ -21,6 +23,10 @@ interface aemenuprops {
 const aemenu = (props: aemenuprops): [ReactNode, (state: boolean, mach: MachineType | null) => void] => {
     let { isDialogOpen, setIsDialogOpen, machineType, setMachineType, refresh} = props;
 
+    const { user } = useAuth();
+    const canCreate = user.permissions.includes(UserPermission.CAN_CREATE_MACHINE_TYPES) || user.permissions.includes(UserPermission.IS_SUPERUSER);
+    const canEdit = user.permissions.includes(UserPermission.CAN_EDIT_MACHINE_TYPES) || user.permissions.includes(UserPermission.IS_SUPERUSER);
+        
     const [name, setName] = useState("");
     const [resourceSlots, setResourceSlots] = useState<ResourceSlot[]>([]);
 
@@ -59,6 +65,10 @@ const aemenu = (props: aemenuprops): [ReactNode, (state: boolean, mach: MachineT
     };
 
     function create() {
+        if (!canCreate) {
+            alert("You do not have the permissions to create machine types");
+            return;
+        }
         if (!name || cost == null) {
             alert("All fields must be populated!");
             return;
@@ -78,6 +88,10 @@ const aemenu = (props: aemenuprops): [ReactNode, (state: boolean, mach: MachineT
 
     function edit() {
         if (!machineType) return;
+        if (!canEdit) {
+            alert("You do not have the permissions to edit machine types");
+            return;
+        }
         if (!name || cost == null) {
             alert("All fields must be populated!");
             return;
@@ -95,52 +109,60 @@ const aemenu = (props: aemenuprops): [ReactNode, (state: boolean, mach: MachineT
         });
     }
 
+    if (!canCreate && !canEdit) return [<></>, () => {}]; 
     return [(
         <Dialog.Root open={isDialogOpen} onOpenChange={handleOpenChange}>
-            <button className="addbtn" onClick={() => { setOpenExtra(true, null); }}><PlusIcon /></button>
-            <Dialog.Portal>
-                <div className='AEdiv'>
-                    <Dialog.Overlay className="DialogOverlay" />
-                    <Dialog.Content className="DialogContent">
-                        <Dialog.Close asChild>
-                            <button className="IconButton" aria-label="Close">
-                                <Cross2Icon />
-                            </button>
-                        </Dialog.Close>
-                        <Dialog.Title className="DialogTitle">{machineType == null ? "Adding" : "Editing"}  Machine Type</Dialog.Title>
-                        <fieldset className="Fieldset">
-                            <label className="Label" htmlFor="name">Name</label>
-                            <input className="Input" id="name" value={name} onChange={e => setName(e.target.value)} maxLength={100} />
+            {canCreate && (
+                <button className="addbtn" onClick={() => { setOpenExtra(true, null); }}><PlusIcon /></button>
+            )}
+            {(canCreate || canEdit) && (
+                <Dialog.Portal>
+                    <div className='AEdiv'>
+                        <Dialog.Overlay className="DialogOverlay" />
+                        <Dialog.Content className="DialogContent">
+                            <Dialog.Close asChild>
+                                <button className="IconButton" aria-label="Close">
+                                    <Cross2Icon />
+                                </button>
+                            </Dialog.Close>
+                            <Dialog.Title className="DialogTitle">{machineType == null ? "Adding" : "Editing"}  Machine Type</Dialog.Title>
+                            <fieldset className="Fieldset">
+                                <label className="Label" htmlFor="name">Name</label>
+                                <input className="Input" id="name" value={name} onChange={e => setName(e.target.value)} maxLength={100} />
 
-                            <label className="Label" htmlFor="resourceSlots">Resource Slots</label>
-                            {resourceSlots.map(resourceSlot => (
-                                <div className="checkbox-labels" key={resourceSlot.id}>
-                                    <input
-                                        className='styled-checkbox'
-                                        type="checkbox"
-                                        checked={resourceSlotIDS.includes(resourceSlot.id)}
-                                        onChange={() => handleCheckboxChange(resourceSlot.id)}
-                                    />
-                                    <label htmlFor='checkbox-layer-shift' className='checkbox-label'>
-                                        {resourceSlot.display_name}
-                                    </label>
-                                </div>
-                            ))}
+                                <label className="Label" htmlFor="resourceSlots">Resource Slots</label>
+                                {resourceSlots.map(resourceSlot => (
+                                    <div className="checkbox-labels" key={resourceSlot.id}>
+                                        <input
+                                            className='styled-checkbox'
+                                            type="checkbox"
+                                            checked={resourceSlotIDS.includes(resourceSlot.id)}
+                                            onChange={() => handleCheckboxChange(resourceSlot.id)}
+                                        />
+                                        <label htmlFor='checkbox-layer-shift' className='checkbox-label'>
+                                            {resourceSlot.display_name}
+                                        </label>
+                                    </div>
+                                ))}
 
-                            <label className="Label" htmlFor="cost">Cost</label>
-                            <input className="Input" placeholder={cost + ""} id="cost" type="number" value={cost} min={0} pattern="[0-9]*" onChange={e => setCost(Math.max(0, e.target.valueAsNumber))} />
-                        </fieldset>
-                        <Dialog.Close asChild>
-                            <button className="Button SaveBtn" onClick={machineType == null ? create : edit}>Save</button>
-                        </Dialog.Close>
-                    </Dialog.Content>
-                </div>
-            </Dialog.Portal>
+                                <label className="Label" htmlFor="cost">Cost</label>
+                                <input className="Input" placeholder={cost + ""} id="cost" type="number" value={cost} min={0} pattern="[0-9]*" onChange={e => setCost(Math.max(0, e.target.valueAsNumber))} />
+                            </fieldset>
+                            <Dialog.Close asChild>
+                                <button className="Button SaveBtn" onClick={machineType == null ? create : edit}>Save</button>
+                            </Dialog.Close>
+                        </Dialog.Content>
+                    </div>
+                </Dialog.Portal>
+            )}
         </Dialog.Root>
     ), setOpenExtra];
 };
 
 const MachineTypes: React.FC = () => {
+    const { user } = useAuth();
+    const canDelete = user.permissions.includes(UserPermission.CAN_DELETE_MACHINE_TYPES) || user.permissions.includes(UserPermission.IS_SUPERUSER);
+    const canEdit = user.permissions.includes(UserPermission.CAN_EDIT_MACHINE_TYPES) || user.permissions.includes(UserPermission.IS_SUPERUSER);
 
     const [data, setData] = React.useState<MachineType[]>([]);
     const columns: (keyof MachineType)[] = data.length > 0 ? (Object.keys(data[0]) as (keyof MachineType)[]).filter( (key) => !key.includes('_id') && key !== 'id' ) : [];
@@ -161,6 +183,10 @@ const MachineTypes: React.FC = () => {
     const refreshPage = () => fetchPage(currentPage);
 
     const onDelete = (index_local: number) => {
+        if (!canDelete) {
+            alert("You do not have the permissions to delete machine types");
+            return;
+        }
         DeleteItem("machinetypes", data[index_local], refreshPage);
     };
 
@@ -180,7 +206,7 @@ const MachineTypes: React.FC = () => {
                 data={data}
                 onDelete={onDelete}
                 onEdit={(e) => setOpen(true, e)}
-                canEdit
+                canEdit={canEdit}
                 currentPage={currentPage}
                 onPageChange={(p) => fetchPage(p)}
                 resourceType={"machinetypes"}
